Fixed line ending problems on the bash scripts
Previous to this commit, some people on windows could not run the bash
scripts because the line endings were unix style.This required people to
run dos2unix or unix2dos on the scripts before being able to run them.

Adding the bash script files to the .gitattributes file means it is
stored with unix style line endings in the remote repo, and git
automatically changes the line endings to the appropriate platform when
the code gets checked out.