Fixed the PID module and added some unit tests to it
The PID module now no longer relies on time. On top of this, it now allows the user the option to provide it with a measured derivative.

A couple of very simple unit tests were added for it and the attitude manager state machine was updated to use its new interface.