Stopped including Hal header files in our unit test build.
This solves all kinds of sketchy problems that arose from including files that were specificaly meant for another hardware architecture. Previous to this, the fix was to compile with -fpermissive and hope for the best, though things started breaking in developement. With this change, the test build will be far more robust. The tradeoff is that we won't be able to unit test any modules that call hal layer functions (like divers).