Ok, PPM catches the interrupt now.
But the biggest changes are to the way I propose we do things in Safety. The way I've set it up, the files that get generated by cubeMX get pulled in by the build system, no more of this copy and paste to other random ass files nonsense. That also means that you can generate whenever and things will still build. And if you want to modify a pin mapping or something, just update the cubeMX file and hit generate, and thats it. The other thing is the code we write will live exclusively in /Src and /Inc. I could see no reason to have the library folder. Now the build system is rigged to not do shit with the stuff in library. Last thing to note is that the main file gets generated by cubeMX (so lives in boardFiles/Src), but we can modify it in specific sections without the next code generation session overwriting those sections.