Built theAutoSteer module (#64)
Built the AutoSteer module.

This module is responsible for converting desired and actual altitude and heading values and turning them into the desired roll, pitch, and rudder angles as well as desired airspeed required to achieve the above heading and altitude.

As of this commit, the implementations are very simplistic. This module has only been designed to function at cruising speed and will not properly takeoff or land. This commit has in mind our first autonomy goal. Past that, this module will have to be revisited.