Safety Code (#77)
* Ok, PPM catches the interrupt now.

But the biggest changes are to the way I propose we do things in Safety. The way I've set it up, the files that get generated by cubeMX get pulled in by the build system, no more of this copy and paste to other random ass files nonsense. That also means that you can generate whenever and things will still build. And if you want to modify a pin mapping or something, just update the cubeMX file and hit generate, and thats it. The other thing is the code we write will live exclusively in /Src and /Inc. I could see no reason to have the library folder. Now the build system is rigged to not do shit with the stuff in library. Last thing to note is that the main file gets generated by cubeMX (so lives in boardFiles/Src), but we can modify it in specific sections without the next code generation session overwriting those sections.

* CubeMX unfortunately only generates a main.c file. So for it to update properly, you have to rename main.cpp to main.c before generation, then back to main.cpp after generation. I've just updated the cmake script to fail if it finds a main.c file, to prevent some confusion

* Nurse PPM manager back in

* remove all the stuff we don't use

* PWM and Safety Controller addition (#68)

* PWM Manager Works

Note: VERIFY TIMER SETTINGS!!

* Add in safety controller

* Fixed the math... I think?

* Fixed timing fr

* Address Changes

Co-authored-by: dakaleyt <32375512+dakaleyt@users.noreply.github.com>

* Safety beginnings ppm (#73)

* PPM changes

using timer to count the gaps,
converting the gaps using the max/min
get_us and get written
not tested

* added the PPM width stuff

might have to move the pulse width stuff to before converting it to us, if accuracy is bad

* Added in Pulse Width and Interchip code

- added pulse width to PPM.cpp
- added in interchip code from safety-nuclear-changes

Co-authored-by: Aaditya <bluebird1339@gmail.com>

* Complete Main Function

* Actual code that Infinion flew with on  first ZeroPilot flight ever (#75)

Co-authored-by: Anthony Berbari <Anthony.berbari@uwaterloo.ca>

* remove IWDG

Co-authored-by: Anthony Berbari <fberbari@edu.uwaterloo.ca>
Co-authored-by: Anthony Berbari <32182266+Fberbari@users.noreply.github.com>
Co-authored-by: dakaleyt <32375512+dakaleyt@users.noreply.github.com>
Co-authored-by: Aaditya Chaudhary <32117362+AadityaChaudhary@users.noreply.github.com>
Co-authored-by: Aaditya <bluebird1339@gmail.com>
Co-authored-by: Anthony Berbari <Anthony.berbari@uwaterloo.ca>