Simulation now also involves Path manager (#117)
The new mechanism for the simulation user to inject data is in AutoPilot/Simulation/Intercepts CommsWithTelemetry_intercept_user.cpp. There, a bunch of PIGO structs can be instantiated as you wish, and each time path manager executes a cycle, it will pick one up.

Now, rather than the simulation data being injected out of the sensor interfaces and going through sensorFusion, the simulation data is injected right out of the sensor fusion interface and into the managers. This way, Simulation can focus on testing the core control algorithms while sensor fusion can be tested independently.

Simulation main is in Autopilot\Simulation\SimDriver\simMain.cpp. That's where the managers are called. Atm attitude Manager is called the same number of times as path manager. If either manager ever enters failed state, a message is displayed and the simulation stops.

Comms between path and attitude managers are done by Autopilot\Simulation\Intercepts\CommsBetweenStateMachines_Intercept.cpp which doesn't involved FREERTOS. As a bonus, every time a communication occurs, the contents of the passed struct are dumped to a file in AutoPilot/SimulationBuild/DebugLogs.

I also had to move a couple things around. Namely, the PIGO struct definition had to be moved from PathDataTypes to TelemPathInterface.hpp. Otherwise, Telemetry manager wouldn't be able to access it. This means I had to add a #include "TelemPathInterface.hpp" to the cruising and landingTakeoff header files.

I gave a namespace to the enums for attitude and path manager statuses because their definitions were interfering with each other.

I had to modify the names of the mail queues between attitude and path and between path and telemetry for the same reason, they were interfering with each other.