Fix FreeRTOS threads
- CubeMX was generating the systick_handler, but was not adding the
osSystickHandler() method to it (which controls the os tick). This
tickhandler was manually added and the HAL_SYSTICK_IRQHandler() which was
there was removed as it did nothing.

- If CubeMX is regenerated it will completely blow this up, so if threads
just stop working for some reason, that is why. Make sure to revert to
the state that this commit is in to fix it